Use BaseChatUITest
