+++ feature/metis/conversation/src/test/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/conversation/BaseChatUITest.kt
@@ -3,12 +3,15 @@ package de.tum.informatics.www1.artemis.native_app.feature.metis.conversation
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.lazy.rememberLazyListState
+import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import de.tum.informatics.www1.artemis.native_app.core.data.DataState
 import de.tum.informatics.www1.artemis.native_app.core.model.Course
 import de.tum.informatics.www1.artemis.native_app.core.model.account.User
 import de.tum.informatics.www1.artemis.native_app.core.test.BaseComposeTest
+import de.tum.informatics.www1.artemis.native_app.core.ui.remote_images.LocalArtemisImageProvider
+import de.tum.informatics.www1.artemis.native_app.core.ui.test.ArtemisImageProviderStub
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.service.MetisModificationFailure
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.service.impl.EmojiServiceStub
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.chatlist.ChatListItem
@@ -16,13 +19,18 @@ import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.chatlist.PostsDataState
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.post.post_actions.PostActionFlags
 import de.tum.informatics.www1.artemis.native_app.feature.metis.conversation.ui.thread.MetisThreadUi
+import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.content.MetisContext
+import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.content.dto.AnswerPost
 import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.content.dto.DisplayPriority
 import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.content.dto.IBasePost
 import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.content.dto.IStandalonePost
+import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.content.dto.StandalonePost
 import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.content.dto.UserRole
 import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.content.dto.conversation.OneToOneChat
 import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.db.pojo.AnswerPostPojo
 import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.db.pojo.PostPojo
+import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.visiblemetiscontextreporter.LocalVisibleMetisContextManager
+import de.tum.informatics.www1.artemis.native_app.feature.metis.shared.visiblemetiscontextreporter.VisiblePostList
 import kotlinx.coroutines.CompletableDeferred
 import kotlinx.coroutines.Deferred
 import kotlinx.datetime.Clock
@@ -33,6 +41,40 @@ abstract class BaseChatUITest : BaseComposeTest() {
 
     private val course: Course = Course(id = 1)
     val conversation = OneToOneChat(id = 2)
+
+    val currentUser = User(
+        id = clientId,
+        name = "Current user"
+    )
+    val otherUser = User(
+        id = 1234,
+        name = "Other user"
+    )
+
+    val simplePostContent = "Simple post content"
+    val simpleAnswerContent = "Simple answer content"
+
+    fun simplePost(
+        postAuthor: User,
+    ): StandalonePost = StandalonePost(
+        id = 1,
+        author = postAuthor,
+        content = simplePostContent,
+    )
+
+    fun simpleThreadPostWithAnswer(
+        postAuthor: User,
+        answerAuthor: User,
+    ): StandalonePost {
+        val basePost = simplePost(postAuthor)
+        val answerPost = AnswerPost(
+            id = 2,
+            author = answerAuthor,
+            content = simpleAnswerContent,
+            post = basePost
+        )
+        return basePost.copy(answers = listOf(answerPost))
+    }
 
     val answers = (0..2).map { index ->
         AnswerPostPojo(
@@ -124,34 +166,44 @@ abstract class BaseChatUITest : BaseComposeTest() {
         onPinPost: (IStandalonePost) -> Deferred<MetisModificationFailure> = { CompletableDeferred() }
     ) {
         composeTestRule.setContent {
-            val list = posts.map { post -> ChatListItem.PostChatListItem(post) }.toMutableList()
-            MetisChatList(
-                modifier = Modifier.fillMaxSize(),
-                initialReplyTextProvider = remember { TestInitialReplyTextProvider() },
-                posts = PostsDataState.Loaded.WithList(list, PostsDataState.NotLoading),
-                clientId = currentUser.id,
-                postActionFlags = PostActionFlags(
-                    isAbleToPin = isAbleToPin,
-                    isAtLeastTutorInCourse = isAtLeastTutorInCourse,
-                    hasModerationRights = hasModerationRights,
-                ),
-                listContentPadding = PaddingValues(),
-                serverUrl = "",
-                courseId = course.id!!,
-                state = rememberLazyListState(),
-                emojiService = EmojiServiceStub,
-                bottomItem = null,
-                isReplyEnabled = true,
-                onCreatePost = { CompletableDeferred() },
-                onEditPost = { _, _ -> CompletableDeferred() },
-                onPinPost = onPinPost,
-                onDeletePost = { CompletableDeferred() },
-                onRequestReactWithEmoji = { _, _, _ -> CompletableDeferred() },
-                onClickViewPost = {},
-                onRequestRetrySend = { _ -> },
-                title = "Title",
-                onFileSelected = { _ -> }
-            )
+            CompositionLocalProvider(
+                LocalArtemisImageProvider provides ArtemisImageProviderStub(),
+                LocalVisibleMetisContextManager provides MockVisibleMetisContextManager.also {
+                    it.registerMetisContext(VisiblePostList(MetisContext.Conversation(
+                        courseId = course.id!!,
+                        conversationId = conversation.id
+                    )))
+                }
+            ) {
+                val list = posts.map { post -> ChatListItem.PostChatListItem(post) }.toMutableList()
+                MetisChatList(
+                    modifier = Modifier.fillMaxSize(),
+                    initialReplyTextProvider = remember { TestInitialReplyTextProvider() },
+                    posts = PostsDataState.Loaded.WithList(list, PostsDataState.NotLoading),
+                    clientId = currentUser.id,
+                    postActionFlags = PostActionFlags(
+                        isAbleToPin = isAbleToPin,
+                        isAtLeastTutorInCourse = isAtLeastTutorInCourse,
+                        hasModerationRights = hasModerationRights,
+                    ),
+                    listContentPadding = PaddingValues(),
+                    serverUrl = "",
+                    courseId = course.id!!,
+                    state = rememberLazyListState(),
+                    emojiService = EmojiServiceStub,
+                    bottomItem = null,
+                    isReplyEnabled = true,
+                    onCreatePost = { CompletableDeferred() },
+                    onEditPost = { _, _ -> CompletableDeferred() },
+                    onPinPost = onPinPost,
+                    onDeletePost = { CompletableDeferred() },
+                    onRequestReactWithEmoji = { _, _, _ -> CompletableDeferred() },
+                    onClickViewPost = {},
+                    onRequestRetrySend = { _ -> },
+                    title = "Title",
+                    onFileSelected = { _ -> }
+                )
+            }
         }
     }
 }
